Remove skins browser and tradeup builder modules
