--- client/src/modules/skins/SkinsBrowser.css
@@ -1,19 +1,0 @@
-.sbc .help { width: 16px; height: 16px; border-radius: 50%; border: 1px solid #2b3b58; display: inline-flex; align-items: center; justify-content: center; font-size: 12px; cursor: help; color: var(--muted); }
-
-.sbc .spinner {
-  width: 16px;
-  height: 16px;
-  border: 2px solid #2b3b58;
-  border-top-color: transparent;
-  border-radius: 50%;
-  animation: sbc-spin 1s linear infinite;
-  margin-right: 8px;
-}
-
-@keyframes sbc-spin {
-  to { transform: rotate(360deg); }
-}
-
-.sbc .toast { padding: 8px 12px; border-radius: 6px; }
-.sbc .toast.success { background: #1d2a44; color: #aaffaa; }
-.sbc .toast.error { background: #401020; color: #ffaaaa; }
