--- client/src/modules/skins/components/StyleTable.css
@@ -1,4 +1,0 @@
-.tableContainer{
-    max-height: 35vh;
-     overflow-y: auto
-}
