--- client/src/modules/skins/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SkinsBrowser";
