--- client/src/modules/skins/services/index.ts
@@ -1,3 +1,0 @@
-export * from "./types";
-export * from "./api";
-export * from "./utils";
