--- client/src/modules/tradeups/TradeupBuilder.css
@@ -1,53 +1,0 @@
-/*
- * Набор утилитарных стилей для TradeupBuilder.tsx. Сетку и адаптивность держим
- * максимально простой, чтобы было удобно дорабатывать внешний вид.
- */
-
-.tradeup-builder {
-  border: 1px solid rgba(255, 255, 255, 0.08);
-}
-
-.tradeup-summary {
-  min-width: 240px;
-}
-
-.tradeup-collections-list {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-}
-
-.tradeup-collections-list .btn {
-  white-space: nowrap;
-}
-
-.tradeup-targets {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
-  gap: 0.75rem;
-}
-
-.tradeup-target-exteriors .btn {
-  min-width: 64px;
-}
-
-.tradeup-hints {
-  display: grid;
-  gap: 0.75rem;
-  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-}
-
-.tradeup-hint ul {
-  padding-left: 1.1rem;
-}
-
-.tradeup-table input,
-.tradeup-table select {
-  min-width: 150px;
-}
-
-.tradeup-results {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(210px, 1fr));
-  gap: 0.75rem;
-}
