+++ client/src/modules/tradeups/types.ts
@@ -1,0 +1,16 @@
+export const EXTERIORS = [
+  "Factory New",
+  "Minimal Wear",
+  "Field-Tested",
+  "Well-Worn",
+  "Battle-Scarred",
+] as const;
+
+export type Exterior = (typeof EXTERIORS)[number];
+
+export interface TradeupInputFormRow {
+  marketHashName: string;
+  collectionId: string;
+  float: string;
+  price: string;
+}
