--- server/src/modules/skins/index.ts
@@ -1,1 +1,0 @@
-export { createSkinsRouter } from "./router";
